+++ docs/assets/js/src/customizer.js
@@ -1,6 +1,7 @@
 /*!
  * Bootstrap Customizer (https://getbootstrap.com/customize/)
  * Copyright 2011-2019 Twitter, Inc.
+ * Copyright 2025 Nozomi Networks (security fixes)
  *
  * Licensed under the Creative Commons Attribution 3.0 Unported License. For
  * details, see https://creativecommons.org/licenses/by/3.0/.
@@ -12,8 +13,9 @@ window.onload = function () { // wait for load in a dumb way because B-0
   'use strict';
 
   var cw = '/*!\n' +
-           ' * Bootstrap v3.4.1 (https://getbootstrap.com/)\n' +
+           ' * Bootstrap v3.4.1-nozomi1 (https://getbootstrap.com/)\n' +
            ' * Copyright 2011-' + new Date().getFullYear() + ' Twitter, Inc.\n' +
+           ' * Copyright 2025 Nozomi Networks (security fixes)\n' +
            ' * Licensed under the MIT license\n' +
            ' */\n\n'