+++ CHANGELOG.md
@@ -1,5 +1,16 @@
+# Changelog
+
+## Nozomi Networks
+
+### v3.4.1-nozomi1 (2025-05-27)
+
+- Fixed CVE-2024-6484 vulnerability in Carousel component that allowed XSS attacks through href and data-target attributes
+- Fixed CVE-2024-6485 vulnerability in Button component
+
+## Upstream
+
 Bootstrap uses [GitHub's Releases feature](https://blog.github.com/2013-07-02-release-your-software/) for its changelogs.
 
-See [the Releases section of our GitHub project](https://github.com/twbs/bootstrap/releases) for changelogs for each release version of Bootstrap.
+See [the Releases section of our GitHub project](https://github.com/NozomiNetworks/bootstrap/releases) for changelogs for each release version of Bootstrap.
 
 Release announcement posts on [the official Bootstrap blog](https://blog.getbootstrap.com/) contain summaries of the most noteworthy changes made in each release.